+++ HallwayGame/src/main/java/vahy/game/config/IGameConfig.java
@@ -2,8 +2,8 @@ package vahy.game.config;
 
 public interface IGameConfig {
 
-    int getDefaultGoalReward();
-    int getDefaultStepPenalty();
+    double getDefaultGoalReward();
+    double getDefaultStepPenalty();
     double getDefaultTrapProbability();
     double getDefaultMissMoveProbability();
 }
